requirements.txt
torch>=2.1.0
transformers>=4.51.0
accelerate>=0.30.0
huggingface_hub>=0.23.0
Pillow>=10.0.0
numpy>=1.24.0
scipy>=1.11.0
scikit-learn>=1.3.0
matplotlib>=3.7.0
seaborn>=0.12.0
pycocotools>=2.0.7
tqdm>=4.65.0
qwen-vl-utils>=0.0.8
